Consolidate bundled plugins: 17 to 12
