--- plugins/changes/herdr-plugin.toml
@@ -1,6 +1,0 @@
-id = "muxr.changes"
-name = "Changes"
-version = "0.1.0"
-min_herdr_version = "0.8.0"
-platforms = ["linux", "macos"]
-description = "Native per-session diff presentation"
