+++ plugins/code/README.md
@@ -1,0 +1,10 @@
+# Code
+
+The repository you're working in, on your phone: files, changes, history, and saved commands.
+
+- **Files:** lazy tree explorer with bounded syntax-highlighted previews (`files.*`).
+- **Changes:** working-tree rows with added/removed counts and app-rendered diffs (`changes.list`).
+- **Git history:** recent commits and per-commit diffs (`history.*`).
+- **Runbook:** saved commands with folder selection (`runbook.*`).
+- **Authority:** read-only except the Runbook's declared write actions. No git mutations.
+- **Removal:** `herdr plugin unlink muxr.code`.
