+++ plugins/code/herdr-plugin.toml
@@ -1,0 +1,6 @@
+id = "muxr.code"
+name = "Code"
+version = "0.1.0"
+min_herdr_version = "0.8.0"
+platforms = ["linux", "macos"]
+description = "Browse files, review changes, read git history, and run saved commands for the repository you are working in"
