--- plugins/file-viewer/herdr-plugin.toml
@@ -1,6 +1,0 @@
-id = "muxr.file-viewer"
-name = "File viewer"
-version = "0.1.0"
-min_herdr_version = "0.8.0"
-platforms = ["linux", "macos"]
-description = "Browse and read files in a git repository"
