--- plugins/file-viewer/muxr-ui.json
@@ -1,126 +1,0 @@
-{
-  "schemaVersion": 1,
-  "pluginId": "muxr.file-viewer",
-  "minMuxrVersion": 11,
-  "contributions": [
-    {
-      "slot": "host.rpc",
-      "id": "repos",
-      "type": "rpc",
-      "method": "repos",
-      "entry": "rpc.mjs",
-      "mode": "read"
-    },
-    {
-      "slot": "host.rpc",
-      "id": "list",
-      "type": "rpc",
-      "method": "list",
-      "entry": "rpc.mjs",
-      "mode": "read"
-    },
-    {
-      "slot": "host.rpc",
-      "id": "read",
-      "type": "rpc",
-      "method": "read",
-      "entry": "rpc.mjs",
-      "mode": "read"
-    },
-    {
-      "slot": "navigation.primary",
-      "id": "nav",
-      "type": "navigation-item",
-      "label": "Files",
-      "icon": "document-text-outline",
-      "contentContributionId": "browse"
-    },
-    {
-      "slot": "navigation.content",
-      "id": "browse",
-      "type": "screen",
-      "title": "{{data.title}}",
-      "data": {
-        "type": "plugin.call",
-        "contributionId": "repos"
-      },
-      "children": [
-        {
-          "type": "list",
-          "title": "Repositories",
-          "emptyText": "No git repositories open",
-          "rows": [],
-          "repeat": {
-            "path": "data.repos",
-            "template": {
-              "type": "row",
-              "title": "{{item.name}}",
-              "subtitle": "{{item.path}}",
-              "action": {
-                "type": "screen",
-                "contributionId": "files",
-                "params": {
-                  "root": "{{item.root}}"
-                }
-              }
-            }
-          }
-        }
-      ]
-    },
-    {
-      "slot": "navigation.content",
-      "id": "files",
-      "type": "screen",
-      "title": "{{data.title}}",
-      "data": {
-        "type": "plugin.call",
-        "contributionId": "list"
-      },
-      "children": [
-        {
-          "type": "text",
-          "text": "{{data.count}}",
-          "tone": "secondary"
-        },
-        {
-          "type": "tree",
-          "title": "Explorer",
-          "emptyText": "No files",
-          "path": "data.tree",
-          "source": { "type": "plugin.call", "contributionId": "list" },
-          "action": {
-            "type": "screen",
-            "contributionId": "file",
-            "params": {
-              "root": "{{data.root}}",
-              "path": "{{item.path}}"
-            }
-          }
-        }
-      ]
-    },
-    {
-      "slot": "navigation.content",
-      "id": "file",
-      "type": "screen",
-      "title": "{{data.name}}",
-      "data": {
-        "type": "plugin.call",
-        "contributionId": "read"
-      },
-      "children": [
-        {
-          "type": "code",
-          "path": "data.body",
-          "fileNamePath": "data.name"
-        },
-        {
-          "type": "text",
-          "text": "{{data.note}}",
-          "tone": "secondary"
-        }
-      ]
-    }
-  ]
-}
