--- plugins/git-history/README.md
@@ -1,9 +1,0 @@
-# Git history
-
-Recent commits for the repository you are working in, and the diff of each.
-
-- **Phone:** a session-header **Git history** button that opens a commit list, then a diff screen.
-- **Host:** read-only `git log` / `git show` RPCs (`log`, `show`).
-- **Authority:** read-only. No writes, no `git` mutations.
-- **Offline:** hidden until the host reconnects.
-- **Removal:** `herdr plugin unlink muxr.git-history`.
