--- plugins/git-history/herdr-plugin.toml
@@ -1,6 +1,0 @@
-id = "muxr.git-history"
-name = "Git history"
-version = "0.1.0"
-min_herdr_version = "0.8.0"
-platforms = ["linux", "macos"]
-description = "Browse recent commits for the repository you are working in, and read each diff"
