--- plugins/ports/herdr-plugin.toml
@@ -1,6 +1,0 @@
-id = "muxr.ports"
-name = "Listening ports"
-version = "0.1.0"
-min_herdr_version = "0.8.0"
-platforms = ["linux", "macos"]
-description = "See what is listening on this machine and stop it from your phone"
