--- plugins/ports/muxr-ui.json
@@ -1,28 +1,0 @@
-{
-  "schemaVersion": 1,
-  "pluginId": "muxr.ports",
-  "minMuxrVersion": 2,
-  "contributions": [
-    { "slot": "host.rpc", "id": "list", "type": "rpc", "method": "list", "entry": "rpc.mjs", "mode": "read" },
-    { "slot": "host.rpc", "id": "detail", "type": "rpc", "method": "detail", "entry": "rpc.mjs", "mode": "read" },
-    { "slot": "host.rpc", "id": "stop", "type": "rpc", "method": "stop", "entry": "rpc.mjs", "mode": "write" },
-    { "slot": "navigation.primary", "id": "nav", "type": "navigation-item", "label": "Ports", "icon": "git-network-outline", "contentContributionId": "browse" },
-    { "slot": "navigation.content", "id": "browse", "type": "screen", "title": "{{data.title}}",
-      "data": { "type": "plugin.call", "contributionId": "list" },
-      "children": [
-        { "type": "list", "title": "Listening", "emptyText": "Nothing listening", "rows": [],
-          "repeat": { "path": "data.ports", "template": {
-            "type": "row", "title": "{{item.title}}", "subtitle": "{{item.subtitle}}",
-            "action": { "type": "screen", "contributionId": "port", "params": { "pid": "{{item.pid}}" } } } } }
-      ]
-    },
-    { "slot": "navigation.content", "id": "port", "type": "screen", "title": "{{data.title}}",
-      "data": { "type": "plugin.call", "contributionId": "detail" },
-      "children": [
-        { "type": "metric", "label": "Process", "value": "{{data.name}}" },
-        { "type": "metric", "label": "PID", "value": "{{data.pid}}" },
-        { "type": "button", "label": "Stop this process", "action": { "type": "plugin.call", "contributionId": "stop" }, "variant": "danger" }
-      ]
-    }
-  ]
-}
