--- plugins/run-server/muxr-ui.json
@@ -1,50 +1,0 @@
-{
-  "schemaVersion": 1,
-  "pluginId": "muxr.run-server",
-  "minMuxrVersion": 5,
-  "contributions": [
-    {
-      "slot": "settings.sections",
-      "id": "run-server-info",
-      "title": "Run Server",
-      "children": [
-        {
-          "type": "row",
-          "title": "Project server",
-          "subtitle": "Starts the selected project's existing dev script for Preview."
-        }
-      ]
-    },
-    {
-      "slot": "session.toolbar",
-      "id": "start-preview-server",
-      "type": "button",
-      "label": "Run server",
-      "action": {
-        "type": "plugin.invoke",
-        "actionId": "start"
-      }
-    },
-    {
-      "slot": "session.header.trailing",
-      "id": "listeners",
-      "type": "native",
-      "primitive": "item-list",
-      "params": {
-        "source": { "type": "plugin.call", "contributionId": "list-listeners" },
-        "title": "Preview",
-        "icon": "globe-outline",
-        "accessibilityLabel": "Open project server",
-        "refreshIntervalMs": 15000
-      }
-    },
-    {
-      "slot": "host.rpc",
-      "id": "list-listeners",
-      "type": "rpc",
-      "method": "list",
-      "entry": "rpc.mjs",
-      "mode": "read"
-    }
-  ]
-}
