--- plugins/runbook/herdr-plugin.toml
@@ -1,6 +1,0 @@
-id = "muxr.runbook"
-name = "Runbook"
-version = "0.1.0"
-min_herdr_version = "0.8.0"
-platforms = ["linux", "macos"]
-description = "Saved commands you can run on your machine from your phone"
