--- plugins/runbook/muxr-ui.json
@@ -1,29 +1,0 @@
-{
-  "schemaVersion": 1,
-  "pluginId": "muxr.runbook",
-  "minMuxrVersion": 10,
-  "contributions": [
-    { "slot": "host.rpc", "id": "list", "type": "rpc", "method": "list", "entry": "rpc.mjs", "mode": "read" },
-    { "slot": "host.rpc", "id": "detail", "type": "rpc", "method": "detail", "entry": "rpc.mjs", "mode": "read" },
-    { "slot": "host.rpc", "id": "run", "type": "rpc", "method": "run", "entry": "rpc.mjs", "mode": "write" },
-    { "slot": "navigation.primary", "id": "nav", "type": "navigation-item", "label": "Runbook", "icon": "play-circle-outline", "contentContributionId": "browse" },
-    { "slot": "navigation.content", "id": "browse", "type": "screen", "title": "{{data.title}}",
-      "data": { "type": "plugin.call", "contributionId": "list" },
-      "children": [
-        { "type": "list", "title": "Saved", "emptyText": "No commands", "rows": [],
-          "repeat": { "path": "data.items", "template": {
-            "type": "row", "title": "{{item.label}}", "subtitle": "{{item.subtitle}}",
-            "action": { "type": "screen", "contributionId": "cmd", "params": { "id": "{{item.id}}" } } } } }
-      ]
-    },
-    { "slot": "navigation.content", "id": "cmd", "type": "screen", "title": "{{data.label}}",
-      "data": { "type": "plugin.call", "contributionId": "detail" },
-      "children": [
-        { "type": "text", "text": "{{data.run}}", "tone": "secondary" },
-        { "type": "field", "kind": "text", "id": "folder", "label": "Run in folder", "value": "{{data.cwd}}" },
-        { "type": "tree", "title": "Choose folder", "emptyText": "No repository folders available", "path": "data.folders", "selectionField": "folder" },
-        { "type": "button", "label": "Run it", "action": { "type": "plugin.call", "contributionId": "run" }, "fields": ["folder"], "variant": "primary" }
-      ]
-    }
-  ]
-}
