+++ plugins/servers/README.md
@@ -1,0 +1,7 @@
+# Servers
+
+What is listening on this machine, and project dev servers for Preview.
+
+- **Ports:** every listener with process and pid; stop stray processes from your phone (declared write).
+- **Run server:** start the selected project's dev script and surface its HTTP listeners in the Preview pill.
+- **Removal:** `herdr plugin unlink muxr.servers`.
