+++ plugins/status/muxr-ui.json
@@ -1,0 +1,120 @@
+{
+  "schemaVersion": 1,
+  "pluginId": "muxr.status",
+  "minMuxrVersion": 13,
+  "contributions": [
+    {
+      "slot": "host.rpc",
+      "id": "usage",
+      "type": "rpc",
+      "method": "usage",
+      "entry": "usage.mjs"
+    },
+    {
+      "slot": "home.cards",
+      "id": "usage.card",
+      "type": "native",
+      "primitive": "item-list",
+      "params": {
+        "title": "Usage",
+        "icon": "speedometer-outline",
+        "accessibilityLabel": "Open coding-agent usage",
+        "source": {
+          "type": "plugin.call",
+          "contributionId": "usage"
+        }
+      }
+    },
+    {
+      "slot": "session.header.trailing",
+      "id": "usage.open",
+      "type": "screen-button",
+      "title": "Usage",
+      "icon": "speedometer-outline",
+      "contentContributionId": "usage.details"
+    },
+    {
+      "slot": "navigation.content",
+      "id": "usage.details",
+      "type": "screen",
+      "title": "Usage",
+      "data": {
+        "type": "plugin.call",
+        "contributionId": "usage"
+      },
+      "children": [
+        {
+          "type": "text",
+          "text": "Local activity only. Costs, prompts, models, projects, and session details stay out.",
+          "tone": "secondary"
+        },
+        {
+          "type": "section",
+          "title": "Today",
+          "columns": 2,
+          "children": [
+            {
+              "type": "metric",
+              "label": "Measured agents",
+              "value": "{{data.summary.measured}}"
+            },
+            {
+              "type": "metric",
+              "label": "Installed agents",
+              "value": "{{data.summary.installed}}"
+            }
+          ]
+        },
+        {
+          "type": "chart",
+          "variant": "bar",
+          "path": "data.activitySeries",
+          "title": "Token activity",
+          "emptyText": "No measured activity today"
+        },
+        {
+          "type": "progress",
+          "path": "data.codexRemaining",
+          "max": 100,
+          "label": "Codex remaining",
+          "valueLabel": "{{data.codexRemainingLabel}}",
+          "tone": "positive"
+        },
+        {
+          "type": "chart",
+          "variant": "ring",
+          "path": "data.codexRing",
+          "title": "Codex availability",
+          "emptyText": "Codex limits unavailable"
+        },
+        {
+          "type": "chart",
+          "variant": "bar",
+          "path": "data.codexSeries",
+          "title": "Codex limits",
+          "emptyText": "Codex limits unavailable"
+        }
+      ]
+    },
+    {
+      "slot": "host.rpc",
+      "id": "vitals",
+      "type": "rpc",
+      "method": "vitals",
+      "entry": "vitals.mjs",
+      "mode": "read"
+    },
+    {
+      "slot": "home.cards",
+      "id": "vitals.card",
+      "type": "data-card",
+      "title": "Machine",
+      "presentation": "sheet",
+      "source": {
+        "type": "plugin.call",
+        "contributionId": "vitals"
+      },
+      "emptyText": "Vitals unavailable"
+    }
+  ]
+}
