--- plugins/usage-status/herdr-plugin.toml
@@ -1,6 +1,0 @@
-id = "muxr.usage-status"
-name = "Usage status"
-version = "0.1.0"
-min_herdr_version = "0.8.0"
-platforms = ["linux", "macos"]
-description = "Native account usage summary"
