--- plugins/vitals/README.md
@@ -1,9 +1,0 @@
-# Machine vitals
-
-Disk, memory, load, and uptime for the machine running your agents.
-
-- **Phone:** a Home **Machine** card (opens as a sheet).
-- **Host:** one read-only `vitals` RPC.
-- **Authority:** read-only. No writes, no shell.
-- **Offline:** hidden until the host reconnects.
-- **Removal:** `herdr plugin unlink muxr.vitals`.
