--- plugins/vitals/herdr-plugin.toml
@@ -1,6 +1,0 @@
-id = "muxr.vitals"
-name = "Machine vitals"
-version = "0.1.0"
-min_herdr_version = "0.8.0"
-platforms = ["linux", "macos"]
-description = "Disk, memory, load and uptime for the machine running your agents"
